# contracts/openapi/payflow-complete.yaml
openapi: 3.1.0
info:
  title: PayFlow API - Advanced Fintech Benchmark
  version: 1.0.0
  description: |
    Comprehensive API for fintech microservices benchmarking.
    Implements advanced patterns: Saga orchestration, CQRS, Event Sourcing,
    idempotency, and distributed transaction management.
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.payflow.example.com
    description: Production server
paths:
...
  /v1/accounts:
    post:
      summary: Create new account
      tags: [Accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          description: Invalid request
        "409":
          description: Account already exists
    
    get:
      summary: List accounts
      tags: [Accounts]
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Accounts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"
...
  /v1/transactions:
    get:
      summary: List transactions
      tags: [Transactions]
      parameters:
        - name: account_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, cancelled]
        - name: type
          in: query
          schema:
            type: string
            enum: [transfer, deposit, withdrawal, fee]
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  
  # Payments
  /v1/payments:
    post:
      summary: Process external payment
      tags: [Payments]
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "201":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          description: Invalid payment request
        "409":
          description: Duplicate payment
...
  /v1/users:
    post:
      summary: Register new user
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Invalid user data
        "409":
          description: User already exists
...
components:
  schemas:
...
    CreateAccountRequest:
      type: object
      required:
        - user_id
        - currency
        - account_type
      properties:
        user_id:
          type: string
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: "USD"
        account_type:
          type: string
          enum: [checking, savings, business]
        initial_balance:
          type: number
          minimum: 0
          default: 0
        metadata:
          type: object
          additionalProperties: true
    AccountResponse:
      type: object
      required:
        - account_id
        - user_id
        - account_number
        - currency
        - status
        - created_at
      properties:
        account_id:
          type: string
        user_id:
          type: string
        account_number:
          type: string
        currency:
          type: string
        account_type:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        balance:
          type: number
        available_balance:
          type: number
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AccountListResponse:
      type: object
      required:
        - accounts
        - pagination
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountResponse"
        pagination:
          $ref: "#/components/schemas/Pagination"
    
    # Transactions
...
    TransactionListResponse:
      type: object
      required:
        - transactions
        - pagination
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionResponse"
        pagination:
          $ref: "#/components/schemas/Pagination"
    TransactionResponse:
      type: object
      required:
        - transaction_id
        - account_id
        - type
        - amount
        - currency
        - status
        - created_at
      properties:
        transaction_id:
          type: string
        account_id:
          type: string
        type:
          type: string
          enum: [transfer, deposit, withdrawal, fee, interest]
        amount:
          type: number
        currency:
          type: string
        balance_after:
          type: number
        reference:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, completed, failed, reversed]
        related_transaction_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    
    # Payments
    CreatePaymentRequest:
      type: object
      required:
        - account_id
        - amount
        - currency
        - payment_method
      properties:
        account_id:
          type: string
        amount:
          type: number
          minimum: 0.01
        currency:
          type: string
          minLength: 3
          maxLength: 3
        payment_method:
          type: string
          enum: [card, bank_transfer, digital_wallet]
        card_details:
          $ref: "#/components/schemas/CardDetails"
        bank_details:
          $ref: "#/components/schemas/BankDetails"
        description:
          type: string
        metadata:
          type: object
          additionalProperties: true
    PaymentResponse:
      type: object
      required:
        - payment_id
        - account_id
        - amount
        - currency
        - status
        - created_at
      properties:
        payment_id:
          type: string
        account_id:
          type: string
        amount:
          type: number
        currency:
          type: string
        payment_method:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        reference:
          type: string
        description:
          type: string
        failure_reason:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    CardDetails:
      type: object
      required:
        - number
        - expiry_month
        - expiry_year
        - cvv
      properties:
        number:
          type: string
        expiry_month:
          type: integer
          minimum: 1
          maximum: 12
        expiry_year:
          type: integer
          minimum: 2024
        cvv:
          type: string
          minLength: 3
          maxLength: 4
        cardholder_name:
          type: string
    BankDetails:
      type: object
      required:
        - account_number
        - routing_number
        - account_holder_name
      properties:
        account_number:
          type: string
        routing_number:
          type: string
        account_holder_name:
          type: string
        bank_name:
          type: string
    
    # Users
    CreateUserRequest:
      type: object
      required:
        - email
        - first_name
        - last_name
        - phone
      properties:
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        date_of_birth:
          type: string
          format: date
        address:
          $ref: "#/components/schemas/Address"
        metadata:
          type: object
          additionalProperties: true
    UserResponse:
      type: object
      required:
        - user_id
        - email
        - first_name
        - last_name
        - status
        - created_at
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
        date_of_birth:
          type: string
          format: date
        address:
          $ref: "#/components/schemas/Address"
        status:
          type: string
          enum: [active, inactive, suspended]
        kyc_status:
          type: string
          enum: [not_started, pending, approved, rejected]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Address:
      type: object
      required:
        - street
        - city
        - state
        - postal_code
        - country
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        country:
          type: string
          minLength: 2
          maxLength: 2
    
    # KYC
...
    Pagination:
      type: object
      required:
        - page
        - limit
        - total
        - pages
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_prev:
          type: boolean
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...